Click handling between message and queue with filters on array proxy content
