+++ addon/components/message-queue.js
@@ -2,9 +2,11 @@
 
 import Em from 'ember';
 import Queue from '../queue';
+import Animations from '../mixins/animations';
 
-export default Em.Component.extend({
-  animationLibrary: 'jQuery',
+export default Em.Component.extend(
+  Animations, {
+
   attributeBindings: ['dataTest:data-test'],
   classNameBindings: ['className'],
   classPrefix: 'flash',
@@ -26,6 +28,12 @@ export default Em.Component.extend({
     return Queue;
   }.property().readOnly(),
 
+  actions: {
+    removeMessage: function(message) {
+      console.log('hellow');
+    }
+  },
+
   show: function() {
     this.setVisibility(true);
   },
@@ -33,39 +41,6 @@ export default Em.Component.extend({
   hide: function() {
     this.setVisibility(false);
   },
-
-  setVisibility: function(shouldShow) {
-    var lib = this.get('animationLibrary');
-    var animationMethod = shouldShow ? 'slideDown' : 'slideUp';
-    var customHideMethod = this.get('customHideMethod');
-    var customShowMethod = this.get('customShowMethod');
-
-    if (this.get('_state') === 'inDOM') {
-      if (shouldShow && customShowMethod) {
-        this.get('customShowMethod')(this);
-      } else if (customHideMethod) {
-        this.get('customHideMethod')(this);
-      } else if (lib === 'jQuery') {
-        this.$()[animationMethod]('fast');
-      } else if (lib === 'velocity') {
-        if (Em.$) {
-          this.$().velocity(animationMethod);
-        } else {
-          Em.run.bind(this, function() {
-            velocity('#' + this.get('elementId'), animationMethod);
-          });
-        }
-      } else {
-        this.set('isVisible', shouldShow);
-      }
-    }
-  },
-
-  _checkForVendor: function() {
-    var lib = this.get('animationLibrary');
-
-    Em.warn('The ' + lib + ' library has not been loaded. Please specify a different value for the animationLibrary option on {{message-queue}}', window[lib]);
-  }.observes('animationLibrary'),
 
   _hideOnLoad: function() {
     var currentMessage = this.get('currentMessage');
