+++ app/templates/components/flash-message.hbs
@@ -12,3 +12,9 @@
   {{!-- Allow block helpers --}}
   {{yield}}
 </dd>
+
+{{#unless timed}}
+  <button>
+    Close
+  </button>
+{{/unless}}
