+++ app/templates/components/message-queue.hbs
@@ -1,5 +1,6 @@
 {{#each nonTimedMessages}}
   {{flash-message
+    action='removeMessage'
     type=type
     content=content
     classPrefix=classPrefix
